move openapi stuff around
