+++ apps/piqueld/src/api/system.rs
@@ -1,7 +1,7 @@
-use axum::{extract::State, http::header, response::IntoResponse};
+use axum::{extract::State, response::IntoResponse};
 use piqueld_client::{SystemCapabilities, SystemStatus};
 
-use super::{ApiState, ok, openapi_document};
+use super::{ApiState, ok};
 
 pub(super) async fn status(State(state): State<ApiState>) -> impl IntoResponse {
     ok(SystemStatus {
@@ -20,14 +20,4 @@ pub(super) async fn capabilities(State(state): State<ApiState>) -> impl IntoResp
         runtime_execution: capabilities.runtime_execution,
         reason: capabilities.reason,
     })
-}
-
-pub(super) async fn openapi() -> impl IntoResponse {
-    (
-        [(
-            header::CONTENT_TYPE,
-            "application/vnd.oai.openapi+json;version=3.1",
-        )],
-        openapi_document().to_string(),
-    )
 }
